Update Delete_Tweet.py
I find some codes online for deleting tweets. I didn't know how to make program reads files and the code I find online for reading file, it was for csv files but twitter archive is .js format so i change it to csv and pull tweet ids from it. If I will learn to read json files I will change to code and so make it available to spesific searchs to delete tweets.